% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataread_monetarypolicy.R
\name{dataread_monetarypolicy}
\alias{dataread_monetarypolicy}
\title{Data for monetary policy analysis.}
\usage{
dataread_monetarypolicy(start.date = "1972-01-01", quarterly = FALSE)
}
\arguments{
\item{start.date}{A string with format 'YYYY-MM-DD'.}

\item{quarterly}{Logical, TRUE averages all data to quarterly frequency.}
}
\value{
A tibble with the data.
}
\description{
Downloads selected data series from FRED (FRB St.Louis), with some rather
minor organization and cleaning.
}
