% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_rbind.R
\name{lapply_rbind}
\alias{lapply_rbind}
\title{Lapply then rbind.}
\usage{
lapply_rbind(X, FUN = function(x) {     x })
}
\arguments{
\item{X}{A list.}

\item{FUN}{A function.}
}
\value{
A matrix or a data.frame, depending on the objects in X.
}
\description{
A wrapper for a do.call('rbind', lapply(X, FUN)).
}
\examples{
lapply_rbind( list(a = c(1,2), b =c(3,4)), sum)
lapply_rbind( as.list(cars), function(x) mean(x))
}
